(maint) Do not let rspec regress into fixtures/modules link
rspec-puppet automagically sets up a symlink in spec/fixtures/modules so
that the assigned RSpec config.modulepath will be able to find module
elements for testing. But the standard rspec spec/**/*_spec.rb filter
will run through the specs via the symlink duplicating test runs unless
excluded. It needs to be excluded at the commandline before rspec
initializes and preps the files (so it's insufficient to set this in the
Config block...)